+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ## 0.2.2 (Unreleased)
 * Excluded actions no longer show up in results (Fixes #106)
+* Credentials Exposure and Service Wildcard now show up in the data file results
 
 ## 0.2.1 (9/25/2020)
 * Fixes issue where Inline Policies were showing up as findings even when they were attached to excluded IAM principals. Fixes #104